+++ ProtoLegacy/EnclaveManifest.cpp
@@ -347,9 +347,9 @@ FloatTupleXYZ EnclaveManifest::SingleToMulti(int input)
 
 	//cout << x << " " << y << " " << z << " " << endl;
 
-	ReturnTuple.x = x;
-	ReturnTuple.y = y;
-	ReturnTuple.z = z;
+	ReturnTuple.x = GLfloat(x);
+	ReturnTuple.y = GLfloat(y);
+	ReturnTuple.z = GLfloat(z);
 
 	return ReturnTuple;
 }
